+++ docs/metrics/extending-the-sdk/MyReader.cs
@@ -27,13 +27,21 @@ internal class MyReader : MetricReader
         this.name = name;
     }
 
-    public override void OnCollect(Batch<Metric> metrics)
+    protected override bool ProcessMetrics(Batch<Metric> metrics, int timeoutMilliseconds)
     {
-        Console.WriteLine($"{this.name}.OnCollect({metrics})");
+        Console.WriteLine($"{this.name}.ProcessMetrics(metrics={metrics}, timeoutMilliseconds={timeoutMilliseconds})");
+        return true;
+    }
+
+    protected override bool OnShutdown(int timeoutMilliseconds)
+    {
+        Console.WriteLine($"{this.name}.OnShutdown(timeoutMilliseconds={timeoutMilliseconds})");
+        return base.OnShutdown(timeoutMilliseconds);
     }
 
     protected override void Dispose(bool disposing)
     {
         Console.WriteLine($"{this.name}.Dispose({disposing})");
+        base.Dispose(disposing);
     }
 }