+++ src/OpenTelemetry.Exporter.Prometheus/PrometheusExporter.cs
@@ -15,7 +15,6 @@
 // </copyright>
 
 using System;
-using System.Collections.Generic;
 using OpenTelemetry.Metrics;
 
 namespace OpenTelemetry.Exporter
@@ -25,10 +24,11 @@ namespace OpenTelemetry.Exporter
     /// </summary>
     [AggregationTemporality(AggregationTemporality.Cumulative)]
     [ExportModes(ExportModes.Pull)]
-    public class PrometheusExporter : BaseExporter<Metric>
+    public class PrometheusExporter : BaseExporter<Metric>, IPullMetricExporter
     {
         internal readonly PrometheusExporterOptions Options;
         internal Batch<Metric> Metrics;
+        private Func<int, bool> funcCollect;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="PrometheusExporter"/> class.
@@ -39,7 +39,11 @@ namespace OpenTelemetry.Exporter
             this.Options = options;
         }
 
-        internal Action CollectMetric { get; set; }
+        public Func<int, bool> Collect
+        {
+            get => this.funcCollect;
+            set { this.funcCollect = value; }
+        }
 
         public override ExportResult Export(in Batch<Metric> metrics)
         {