+++ src/OpenTelemetry/Metrics/BaseExportingMetricReader.cs
@@ -15,15 +15,16 @@
 // </copyright>
 
 using System;
+using System.Diagnostics;
+using System.Threading;
 
 namespace OpenTelemetry.Metrics
 {
     public class BaseExportingMetricReader : MetricReader
     {
         protected readonly BaseExporter<Metric> exporter;
-        protected bool disposed;
-
         private readonly ExportModes supportedExportModes = ExportModes.Push | ExportModes.Pull;
+        private bool disposed;
 
         public BaseExportingMetricReader(BaseExporter<Metric> exporter)
         {
@@ -44,14 +45,27 @@ namespace OpenTelemetry.Metrics
                 var attr = (ExportModesAttribute)attributes[attributes.Length - 1];
                 this.supportedExportModes = attr.Supported;
             }
+
+            if (exporter is IPullMetricExporter pullExporter)
+            {
+                if (this.supportedExportModes.HasFlag(ExportModes.Push))
+                {
+                    pullExporter.Collect = this.Collect;
+                }
+                else
+                {
+                    pullExporter.Collect = (timeoutMilliseconds) =>
+                    {
+                        using (PullMetricScope.Begin())
+                        {
+                            return this.Collect(timeoutMilliseconds);
+                        }
+                    };
+                }
+            }
         }
 
         protected ExportModes SupportedExportModes => this.supportedExportModes;
-
-        public override void OnCollect(Batch<Metric> metrics)
-        {
-            this.exporter.Export(metrics);
-        }
 
         internal override void SetParentProvider(BaseProvider parentProvider)
         {
@@ -60,23 +74,77 @@ namespace OpenTelemetry.Metrics
         }
 
         /// <inheritdoc/>
+        protected override bool ProcessMetrics(Batch<Metric> metrics, int timeoutMilliseconds)
+        {
+            return this.exporter.Export(metrics) == ExportResult.Success;
+        }
+
+        /// <inheritdoc />
+        protected override bool OnCollect(int timeoutMilliseconds)
+        {
+            if (this.supportedExportModes.HasFlag(ExportModes.Push))
+            {
+                return base.OnCollect(timeoutMilliseconds);
+            }
+
+            if (this.supportedExportModes.HasFlag(ExportModes.Pull) && PullMetricScope.IsPullAllowed)
+            {
+                return base.OnCollect(timeoutMilliseconds);
+            }
+
+            // TODO: add some error log
+            return false;
+        }
+
+        /// <inheritdoc />
+        protected override bool OnShutdown(int timeoutMilliseconds)
+        {
+            var result = true;
+
+            if (timeoutMilliseconds == Timeout.Infinite)
+            {
+                result = this.Collect(Timeout.Infinite) && result;
+                result = this.exporter.Shutdown(Timeout.Infinite) && result;
+            }
+            else
+            {
+                var sw = Stopwatch.StartNew();
+                result = this.Collect(timeoutMilliseconds) && result;
+                var timeout = timeoutMilliseconds - sw.ElapsedMilliseconds;
+                result = this.exporter.Shutdown((int)Math.Max(timeout, 0)) && result;
+            }
+
+            return result;
+        }
+
+        /// <inheritdoc/>
         protected override void Dispose(bool disposing)
         {
-            base.Dispose(disposing);
+            if (this.disposed)
+            {
+                return;
+            }
 
-            if (disposing && !this.disposed)
+            if (disposing)
             {
                 try
                 {
+                    if (this.exporter is IPullMetricExporter pullExporter)
+                    {
+                        pullExporter.Collect = null;
+                    }
+
                     this.exporter.Dispose();
                 }
                 catch (Exception)
                 {
                     // TODO: Log
                 }
-
-                this.disposed = true;
             }
+
+            this.disposed = true;
+
+            base.Dispose(disposing);
         }
     }
 }