+++ test/OpenTelemetry.Tests/Metrics/MetricAPITest.cs
@@ -39,6 +39,60 @@ namespace OpenTelemetry.Metrics.Tests
         }
 
         [Theory]
+        [InlineData(AggregationTemporality.Cumulative)]
+        [InlineData(AggregationTemporality.Delta)]
+        public void StreamNamesDuplicatesAreNotAllowedTest(AggregationTemporality temporality)
+        {
+            var metricItems = new List<Metric>();
+            int metricCount = 0;
+            var metricExporter = new TestExporter<Metric>(ProcessExport);
+
+            void ProcessExport(Batch<Metric> batch)
+            {
+                foreach (var metric in batch)
+                {
+                    metricCount++;
+                }
+            }
+
+            var metricReader = new BaseExportingMetricReader(metricExporter)
+            {
+                PreferredAggregationTemporality = temporality,
+            };
+            using var meter1 = new Meter("TestDuplicateMetricName1");
+            using var meter2 = new Meter("TestDuplicateMetricName2");
+            using var meterProvider = Sdk.CreateMeterProviderBuilder()
+                .AddSource("TestDuplicateMetricName1")
+                .AddSource("TestDuplicateMetricName2")
+                .AddReader(metricReader)
+                .Build();
+
+            // Expecting one metric stream.
+            var counterLong = meter1.CreateCounter<long>("name1");
+            counterLong.Add(10);
+            metricReader.Collect();
+            Assert.Equal(1, metricCount);
+
+            // The following will be ignored as
+            // metric of same name exists.
+            // Metric stream will remain one.
+            var anotherCounterSameName = meter1.CreateCounter<long>("name1");
+            anotherCounterSameName.Add(10);
+            metricCount = 0;
+            metricReader.Collect();
+            Assert.Equal(1, metricCount);
+
+            // The following will also be ignored
+            // as the name is same.
+            // (the Meter name is not part of stream name)
+            var anotherCounterSameNameDiffMeter = meter2.CreateCounter<long>("name1");
+            anotherCounterSameNameDiffMeter.Add(10);
+            metricCount = 0;
+            metricReader.Collect();
+            Assert.Equal(1, metricCount);
+        }
+
+        [Theory]
         [InlineData(true)]
         [InlineData(false)]
         public void CounterAggregationTest(bool exportDelta)
@@ -59,11 +113,11 @@ namespace OpenTelemetry.Metrics.Tests
                 PreferredAggregationTemporality = exportDelta ? AggregationTemporality.Delta : AggregationTemporality.Cumulative,
             };
 
-            var meter = new Meter("TestMeter");
+            using var meter = new Meter("TestMeter");
             var counterLong = meter.CreateCounter<long>("mycounter");
-            var meterProvider = Sdk.CreateMeterProviderBuilder()
+            using var meterProvider = Sdk.CreateMeterProviderBuilder()
                 .AddSource("TestMeter")
-                .AddMetricReader(metricReader)
+                .AddReader(metricReader)
                 .Build();
 
             counterLong.Add(10);
@@ -135,7 +189,7 @@ namespace OpenTelemetry.Metrics.Tests
                 PreferredAggregationTemporality = exportDelta ? AggregationTemporality.Delta : AggregationTemporality.Cumulative,
             };
 
-            var meter = new Meter(meterName);
+            using var meter = new Meter(meterName);
             int i = 1;
             var counterLong = meter.CreateObservableCounter<long>(
             "observable-counter",
@@ -146,9 +200,9 @@ namespace OpenTelemetry.Metrics.Tests
                     new Measurement<long>(i++ * 10),
                 };
             });
-            var meterProvider = Sdk.CreateMeterProviderBuilder()
+            using var meterProvider = Sdk.CreateMeterProviderBuilder()
                 .AddSource(meterName)
-                .AddMetricReader(metricReader)
+                .AddReader(metricReader)
                 .Build();
 
             metricReader.Collect();
@@ -193,7 +247,7 @@ namespace OpenTelemetry.Metrics.Tests
             {
                 foreach (var metric in batch)
                 {
-                    foreach (var metricPoint in metric.GetMetricPoints())
+                    foreach (ref var metricPoint in metric.GetMetricPoints())
                     {
                         metricPointCount++;
                     }
@@ -204,11 +258,11 @@ namespace OpenTelemetry.Metrics.Tests
             {
                 PreferredAggregationTemporality = temporality,
             };
-            var meter = new Meter("TestMeter");
-            var counterLong = meter.CreateCounter<long>("mycounter");
-            var meterProvider = Sdk.CreateMeterProviderBuilder()
-                .AddSource("TestMeter")
-                .AddMetricReader(metricReader)
+            using var meter = new Meter("TestPointCapMeter");
+            var counterLong = meter.CreateCounter<long>("mycounterCapTest");
+            using var meterProvider = Sdk.CreateMeterProviderBuilder()
+                .AddSource("TestPointCapMeter")
+                .AddReader(metricReader)
                 .Build();
 
             // Make one Add with no tags.
@@ -256,11 +310,11 @@ namespace OpenTelemetry.Metrics.Tests
                 PreferredAggregationTemporality = AggregationTemporality.Cumulative,
             };
 
-            var meter = new Meter("TestMeter");
+            using var meter = new Meter("TestLongCounterMeter");
             var counterLong = meter.CreateCounter<long>("mycounter");
-            var meterProvider = Sdk.CreateMeterProviderBuilder()
-                .AddSource("TestMeter")
-                .AddMetricReader(metricReader)
+            using var meterProvider = Sdk.CreateMeterProviderBuilder()
+                .AddSource("TestLongCounterMeter")
+                .AddReader(metricReader)
                 .Build();
 
             // setup args to threads.
@@ -284,8 +338,7 @@ namespace OpenTelemetry.Metrics.Tests
             // Block until all threads started.
             mreToEnsureAllThreadsStarted.WaitOne();
 
-            Stopwatch sw = new Stopwatch();
-            sw.Start();
+            Stopwatch sw = Stopwatch.StartNew();
 
             // unblock all the threads.
             // (i.e let them start counter.Add)
@@ -300,7 +353,6 @@ namespace OpenTelemetry.Metrics.Tests
             var timeTakenInMilliseconds = sw.ElapsedMilliseconds;
             this.output.WriteLine($"Took {timeTakenInMilliseconds} msecs. Total threads: {numberOfThreads}, each thread doing {numberOfMetricUpdateByEachThread} recordings.");
 
-            meterProvider.Dispose();
             metricReader.Collect();
 
             var sumReceived = GetLongSum(metricItems);
@@ -327,11 +379,11 @@ namespace OpenTelemetry.Metrics.Tests
                 PreferredAggregationTemporality = AggregationTemporality.Cumulative,
             };
 
-            var meter = new Meter("TestMeter");
+            using var meter = new Meter("TestDoubleCounterMeter");
             var counterDouble = meter.CreateCounter<double>("mycounter");
-            var meterProvider = Sdk.CreateMeterProviderBuilder()
-                .AddSource("TestMeter")
-                .AddMetricReader(metricReader)
+            using var meterProvider = Sdk.CreateMeterProviderBuilder()
+                .AddSource("TestDoubleCounterMeter")
+                .AddReader(metricReader)
                 .Build();
 
             // setup args to threads.
@@ -355,8 +407,7 @@ namespace OpenTelemetry.Metrics.Tests
             // Block until all threads started.
             mreToEnsureAllThreadsStarted.WaitOne();
 
-            Stopwatch sw = new Stopwatch();
-            sw.Start();
+            Stopwatch sw = Stopwatch.StartNew();
 
             // unblock all the threads.
             // (i.e let them start counter.Add)
@@ -371,7 +422,6 @@ namespace OpenTelemetry.Metrics.Tests
             var timeTakenInMilliseconds = sw.ElapsedMilliseconds;
             this.output.WriteLine($"Took {timeTakenInMilliseconds} msecs. Total threads: {numberOfThreads}, each thread doing {numberOfMetricUpdateByEachThread} recordings.");
 
-            meterProvider.Dispose();
             metricReader.Collect();
 
             var sumReceived = GetDoubleSum(metricItems);
@@ -385,7 +435,7 @@ namespace OpenTelemetry.Metrics.Tests
             long sum = 0;
             foreach (var metric in metrics)
             {
-                foreach (var metricPoint in metric.GetMetricPoints())
+                foreach (ref var metricPoint in metric.GetMetricPoints())
                 {
                     sum += metricPoint.LongValue;
                 }
@@ -399,7 +449,7 @@ namespace OpenTelemetry.Metrics.Tests
             double sum = 0;
             foreach (var metric in metrics)
             {
-                foreach (var metricPoint in metric.GetMetricPoints())
+                foreach (ref var metricPoint in metric.GetMetricPoints())
                 {
                     sum += metricPoint.DoubleValue;
                 }
@@ -420,7 +470,7 @@ namespace OpenTelemetry.Metrics.Tests
             var mre = arguments.MreToBlockUpdateThread;
             var mreToEnsureAllThreadsStart = arguments.MreToEnsureAllThreadsStart;
             var counter = arguments.Counter;
-
+            var valueToUpdate = arguments.DeltaValueUpdatedByEachCall;
             if (Interlocked.Increment(ref arguments.ThreadsStartedCount) == numberOfThreads)
             {
                 mreToEnsureAllThreadsStart.Set();
@@ -431,7 +481,7 @@ namespace OpenTelemetry.Metrics.Tests
 
             for (int i = 0; i < numberOfMetricUpdateByEachThread; i++)
             {
-                counter.Add(arguments.DeltaValueUpdatedByEachCall, new KeyValuePair<string, object>("verb", "GET"));
+                counter.Add(valueToUpdate, new KeyValuePair<string, object>("verb", "GET"));
             }
         }